Refactor using partial of NpmRunner to simplify usage syntax
