+++ src/Cake.Npm/NpmRunner.cs
@@ -9,9 +9,10 @@ namespace Cake.Npm
     /// <summary>
     /// A wrapper around the Node Npm package manager
     /// </summary>
-    public class NpmRunner : Tool<NpmRunnerSettings>
+    public partial class NpmRunner : Tool<NpmRunnerSettings>
     {
-        /// <summary>
+
+	    /// <summary>
         /// Initializes a new instance of the <see cref="NpmRunner" /> class.
         /// </summary>
         /// <param name="fileSystem">The file system</param>
@@ -19,10 +20,10 @@ namespace Cake.Npm
         /// <param name="processRunner">The process runner</param>
         /// <param name="globber">The globber</param>
         internal NpmRunner(IFileSystem fileSystem, ICakeEnvironment environment, IProcessRunner processRunner, IGlobber globber) : base(fileSystem, environment, processRunner, globber)
-        {
-        }
-        
-        /// <summary>
+	    {
+	    }
+
+	    /// <summary>
         /// execute 'npm install' with options
         /// </summary>
         /// <param name="configure">options when running 'npm install'</param>
@@ -64,7 +65,7 @@ namespace Cake.Npm
             return args;
         }
 
-        /// <summary>
+	    /// <summary>
         /// Gets the name of the tool
         /// </summary>
         /// <returns>the name of the tool</returns>
@@ -84,3 +85,4 @@ namespace Cake.Npm
         }
     }
 }
+
